Correção de erros e regras de negócio em reservas, empréstimos e devoluções.

+++ database/migrations/2026_12_01_021453_create_emprestimos_table.php
@@ -16,7 +16,7 @@ class CreateEmprestimosTable extends Migration
         Schema::create('emprestimos', function (Blueprint $table) {
             $table->id();
             $table->timestamps();            
-            $table->decimal('quantidadeemprestada', 6,2);
+            $table->integer('quantidadeemprestada');
             $table->date('dataemprestimo');  
             $table->unsignedBigInteger('usuario_id');
             $table->foreign('usuario_id')->references('id')->on('users');
